+++ signalwire/relay/calling/results/stop_result.py
@@ -1,0 +1,3 @@
+class StopResult:
+  def __init__(self, successful):
+    self.successful = successful