+++ signalwire/tests/relay/calling/test_call.py
@@ -2,12 +2,6 @@ import asyncio
 import json
 import pytest
 from unittest.mock import Mock
-from signalwire.relay.calling import Call
-
-@pytest.fixture()
-def relay_call(relay_calling):
-  params = json.loads('{"call_state":"created","context":"office","device":{"type":"phone","params":{"from_number":"+12029999999","to_number":"+12028888888"}},"direction":"inbound","call_id":"call-id","node_id":"node-id","tag":"call-tag"}')
-  return Call(calling=relay_calling, **params)
 
 def test_init_options(relay_call):
   assert relay_call.id == 'call-id'