+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased]
 ### Added
 - Support for Relay Messaging
+- Support for Calling `connect` and `play` methods.
 
 ## [2.0.0b2] - 2019-10-03
 ### Fixed